+++ src/panels/config/energy/components/ha-energy-grid-settings.ts
@@ -40,7 +40,7 @@ import {
   showEnergySettingsGridFlowFromDialog,
   showEnergySettingsGridFlowToDialog,
 } from "../dialogs/show-dialogs-energy";
-import "./ha-energy-validation-message";
+import "./ha-energy-validation-result";
 import { energyCardStyles } from "./styles";
 
 @customElement("ha-energy-grid-settings")
