+++ src/panels/config/energy/components/ha-energy-validation-result.ts
