+++ src/translations/en.json
@@ -1111,7 +1111,7 @@
                 "title": "Unexpected unit of measurement",
                 "description": "The following entities do not have expected units of measurement that ends with /kWh or /Wh:"
               },
-              "entity_unexpected_state_class": {
+              "entity_unexpected_state_class_total_increasing": {
                 "title": "Unexpected state class",
                 "description": "The following entities do not have expeted state class \"total_increasing\""
               }
